[FLINK-40552][runtime] Make asynchronous transform function self-contained
